refactor: move village codes schema into separate file
