+++ src/components/village-codes/ChangeModal.tsx
@@ -1,7 +1,7 @@
 import { trpc } from "@/utils/trpc";
 import toast from "react-hot-toast";
 import Modal from "@/components/interactive/Modal";
-import { NewVillageCode } from "@/db/schema/schema";
+import { NewVillageCode } from "@/db/schema/villageCodes";
 import { DEFAULT_FORM, FormFields } from "@/lib/utils/villageCodeTypes";
 import {
   FormProvider,
