+++ src/db/schema/schema.ts
@@ -11,6 +11,8 @@ import {
   uuid,
 } from "drizzle-orm/pg-core";
 
+import { villageCodes } from "./villageCodes";
+
 // Stub for Supabase-managed auth schema — not migrated, used only for FK references
 const authSchema = pgSchema("auth");
 export const authUsers = authSchema.table("users", {
@@ -21,32 +23,6 @@ export const authUsers = authSchema.table("users", {
 export const genderEnum = pgEnum("gender", ["male", "female"]);
 
 // Define tables
-
-/*
-Village Codes Table:
-- id: Primary key, auto-incrementing integer.
-- code: Unique code for the village (e.g., "V001").
-- name: Name of the village (e.g., "Village A").
-- colorHex: Hexadecimal color code for the village (e.g., "#FF5733").
-- isVisible: Boolean indicating whether the village code is visible in the application.
-- createdAt: Timestamp of when the record was created.
-- updatedAt: Timestamp of when the record was last updated.   
-*/
-export const villageCodes = pgTable("village_codes", {
-  id: serial("id").primaryKey(),
-  code: varchar("code", { length: 255 }).notNull().unique(),
-  name: varchar("name", { length: 255 }).notNull(),
-  colorHex: varchar("color_hex", { length: 7 }).notNull(), // maps to 'color_hex' in DB
-  isVisible: boolean("is_visible").default(true).notNull(), // maps to 'is_visible' in DB
-  createdAt: timestamp("created_at").defaultNow().notNull(),
-  updatedAt: timestamp("updated_at")
-    .defaultNow()
-    .notNull()
-    .$onUpdate(() => new Date()),
-});
-
-export type VillageCode = typeof villageCodes.$inferSelect;
-export type NewVillageCode = typeof villageCodes.$inferInsert;
 
 /*
 Patients Table:
