+++ src/db/schema/villageCodes.ts
@@ -1,0 +1,33 @@
+import {
+  pgTable,
+  serial,
+  varchar,
+  boolean,
+  timestamp,
+} from "drizzle-orm/pg-core";
+
+/*
+Village Codes Table:
+- id: Primary key, auto-incrementing integer.
+- code: Unique code for the village (e.g., "V001").
+- name: Name of the village (e.g., "Village A").
+- colorHex: Hexadecimal color code for the village (e.g., "#FF5733").
+- isVisible: Boolean indicating whether the village code is visible in the application.
+- createdAt: Timestamp of when the record was created.
+- updatedAt: Timestamp of when the record was last updated.   
+*/
+export const villageCodes = pgTable("village_codes", {
+  id: serial("id").primaryKey(),
+  code: varchar("code", { length: 255 }).notNull().unique(),
+  name: varchar("name", { length: 255 }).notNull(),
+  colorHex: varchar("color_hex", { length: 7 }).notNull(), // maps to 'color_hex' in DB
+  isVisible: boolean("is_visible").default(true).notNull(), // maps to 'is_visible' in DB
+  createdAt: timestamp("created_at").defaultNow().notNull(),
+  updatedAt: timestamp("updated_at")
+    .defaultNow()
+    .notNull()
+    .$onUpdate(() => new Date()),
+});
+
+export type VillageCode = typeof villageCodes.$inferSelect;
+export type NewVillageCode = typeof villageCodes.$inferInsert;
