+++ src/lib/utils/villageCodeTypes.ts
@@ -1,4 +1,4 @@
-import { NewVillageCode } from "@/db/schema/schema";
+import { NewVillageCode } from "@/db/schema/villageCodes";
 
 export const DEFAULT_FORM: NewVillageCode = {
   code: "",
